requirements.txt
requests==2.25.1
tensorflow>=2.3.0
streamlit==0.76.0
google_api_python_client==1.12.8
protobuf==3.14.0
opencv-python
numpy
matplotlib
object_detection
Flask